Refactor WebHookController: Extracted business logic into WebHookService for handling payment events.
